TWO SUM PROBLEM OF LEETCODE
Intuition :-
This problem can easily be solved using Map in JavaScript. Let's see how.

Approach :-
We iterate each item in the nums array and check if (target-current_value) is present in the Map.

If not present then set it as key and its index as value.

If present push the current index and value of the found key(which is an index) in an array. Now return the array.